chore(lib): undo changes from eda48c5ce0cda2afe3858f20db75e858b9f8a4b3

+++ src/rt/utils.rs
@@ -17,16 +17,16 @@ pub fn generate_all_path_candidates(
         // Zero path of size order
         return Array2::default((order as usize, 0)).into_pyarray(py);
     } else if order == 1 {
-        let mut path_candidates = Array2::default((1, num_primitives));
+        let mut path_candidates = Array2::default((1, num_primitives as usize));
 
         for j in 0..num_primitives {
-            path_candidates[(0, j)] = j;
+            path_candidates[(0, j as usize)] = j;
         }
         return path_candidates.into_pyarray(py);
     }
-    let num_choices = num_primitives - 1;
+    let num_choices = (num_primitives - 1) as usize;
     let num_candidates_per_batch = num_choices.pow(order - 1);
-    let num_candidates = num_primitives * num_candidates_per_batch;
+    let num_candidates = (num_primitives as usize) * num_candidates_per_batch;
 
     let mut path_candidates = Array2::default((order as usize, num_candidates));
     let mut batch_size = num_candidates_per_batch;
